GSL quantile function

+++ src/cross-level/stats-cl.h
@@ -67,7 +67,7 @@ float tscore_Norm_z(float p);
 float tscore_Hills_inv_t(float p, int idf);
 float tscore_T_z(float t, int df);
 float tscore_T_p(float t, int df);
-float quantile(float *x, int n, float p);
+float quantile(double *x, int n, float p);
 int mode(int *x, int n);
 int n_uniq(int *x, int n);
 int **histogram(int *x, int n, int *n_uniq);
